README: replace broken link
The link to the DemoSnapshotTests.swift file was broken due to the file
being removed in 4dcd04f4605a (Deletes conflicting file, 2018-09-04).
Not sure which file is the most appropriate to use but for now just
changed it to the test directory.

Also while at it use a link for the FullscreenViewTests reference.