Create a new `LoginEntryView`
One of the considerations for this view, specially for
`LoginEntryDialogueView` is that the title `Label`'s font will be
overriten by a local `title2Strong` font as that style is not yet into
the defined styles within the design system. Nonetheless is a much
better fit for this view than just using the existing `title2` style.