Set dimensions for arrow view in settings view
By default the assets is a bit bigger than expected, since the asset may
be used in other places I opted for setting the dimensions specifically
for this use case.

Then, this whole settings view is used as a way to customize the default
cell appearance to be used in the setting screen and modify certain
internal cell aspects like the separator.