Initial implementation of the Klima Brølet view
This is a new campaign at the style of Earth Hour.

I initially copied and modified the earth hour views and adapted them to
this current use case, without the custom header animation.

One of the extensions from this view in relation to the Earth Hour one
is the support for dynamic type and scrollable content.

An extension to this view is to make it a more reusable component as a
campaign card component. The main reason I didn't do it right now is
because time-to-ship is fairly limited, together with I am not sure how
to solve the potential injection of a view as the scrollable content of
the campaign card.