Favor setting constraint constants over using margins guide
Creating constraints towards the `layoutMarginGuide` reads better, but
the behavior is different for iOS 10 (I am not sure exactly why, it's
not mentioned in the docs). The result in iOS 10 would be like to set a
constraint to the actual border of the view, not considering its
margins.

Then to ensure the layout works as intended for all of our supported iOS
versions, I set the constants instead.