Add basic dark mode snaphshots
There is a strange issue with dark mode + test + snapshot testing, as some part
of the UI react to dark mode, but not others.

Our current theory is that: elements that are not set custom colors and
just use system ones get rendered in dark mode correctly, but not
elements with custom finn colors (even though they support dark mode).

This can potentially be becasue when applying dark mode there is an
animation involved and that animation has not finished when the snapshot
is taken.

I will be debugging this.