Add `ChristmasWishListView` component
This is a view that will be rendered as a pop up, meaning the view
itself will have constraints that limit its width and heith.

The view implementation includes the subview:

- ChristmasWishListView.ContentView: a nested subview to display the
text content of the popup
- ChristmasWishListView.ChrismasButton: a custom call-to-action like
button with custom colors
- MissionPopupView: a new generic popup-like container view. The reason
for this is that pop up with this structure are not that infrequent,
then I tried to make it a bit more generic, although for now I leave the
subview within the Christmas folder until it's really needed as a
generic component.

The `ChristmasWishListView` implementation sets the contentView  of the
`MissionPopupView` as a UIView that contains two instances of
`ContentView` on top of each other. Then with a `CGAffineTransform` the
second page content will be translated to the right to give the page
behavior when clicking the button.

I felt this was the easiest approach while being able to create a more
reusable component in `MissionPopupView`.