Only show one broadcast for the demo (and snapshot)
We were showing 2 broadcast as sample for the BroadcastDemoView. Then,
since they are defined in a set, we don't have a determined order of
elements, so after merging #694 (the test passed), master failed since
the test was non-deterministic.

The easy way to ensure the test is deterministic, while keeping the
correct behavior of using a `Set` for the `Broadcast` view, is to just
show one sample then.