Refactor:`Button.Style` as a `struct` instead of an `enum` (#723)
* Re-write Button.Style as struct instead of an enum

The reason for this is that it's not possible to override or exten
button styles easily as we cannot add new cases to `enums` in
extensions.

Sometimes, we may want a call to action button with a different color.
Then we may attemp to set the `backgroundColor` of the button, but then
when pressing the button the `isHighlighted` override will set the
`highlightedBodyColor` instead, resulting in an unexpected behavior.

Then I decided to go for an Struct approach, as adding new cases is as
simple as adding a new static value and from the API side, the
`Button`/`Style` API doens't change at all.

When creating the default button styles you can see a bit of
duplication in terms of defining all properties (even the common ones)
for each single style. I take this as a good thing, as now we can see
all the settings for a given style in one place, as oppose to traverse a
bunch of switch statements to figure out the actual values of a given
style.

* Add `overrideStyle` method to `Button.Style`

The idea is to be able to create new styles from existing ones while
overriding part of the properties only.

This is intended to be used for styles created for a concrete usage
only, as for default styles, it's better to list all properties to avoid
unintended changes when changing another default style

* Add `ButtonStateStyle`

As a convenience structure to define styles as properties like:

 - highlightedBodyColor
 - highlightedTextColor
 - highlightedBorderColor

are related, but the extra repetition here impacts readability.

Underneath the `Button.Style` still uses those properties for easy
access to them within the `Button` code.

* Refactor default button styles with the new initializer

* Add initializer that takes `[UIControl.State: StateStyle]`

As a convenience to define button styles in a nicer way.

I'm keeping the flat properties for easy of access and also to define
default values for those properties.

It's a compromise in easy to define styles and easy to use them
internally.