Allow setting `accentColor` for `StepSlider`
This is to modify the active range color as well of the thumb image.

In the case of the thumb image, since colors are set on runtime, I need
to create a new image for the accent color, for the normal and
highlighted states.

For dark mode, the accent color will resolve with a different value, so
then the thumb image will also need to be regenerated to show this
change in the value, since the image is generated in code it doesn't
have a dark mode counterpart.

If the `accentColor` is never set, the slider will use the default
colors/images as its has been doing so far.