Expose UIColor extra initializers / convenience methods
These are very common UIColor patterns that we need across the code that
depends on FinniversKit, then it's nice just to provide that
functionality in it.

Colors usually come as semantic colors directly from FinniversKit, but
in PRs like #724 where we will need to handle colors from the backend
(as hex values), it's useful to have this functionality from
FinniversKit.

There are two similar methods to notice here:

- `dynamicColor`
- `dynamicColorIfAvailable`

This functionality is quite common as well in cases were multiple iOS
version are supported. The first method just takes into account the iOS
capabilities. The second method also considers the settings in
FinniversKit to globally enable/disable dark mode for its components.