Ensure labels are part of the `HappinessRatingView`
I noticed that originally the `HappinessRatingView` when being centered on
the Y axis didn't consider the labels on the alignment, resulting in an slightly
off centering which was done solely for the rating faces.

The reason for that is that the labels were outside the bounds of the stack view
as the stack view contained images and the label was underneath the images.

Then I replaced the stack of image views for a stack of rating views, which
in turn are StackViews that contains an image view and a label. Then, if the labels
are not needed, they won't be added to the stack view, meaning the RatingView
stackView will be as big as the iamge alone.