Improve the API for CollapsibleContentDemoView
I did a sample integration with the object page and for that case, we needed
to control the expansion of the content manually, so i added that possibility
and check all the features of the component needed for the object page are
present.

I also ensured that when the layout margins of the view are modified, the
constant that determine the inner positions are updated.

Finally i noticed in the demo view some touches were not responsive in certain
areas of the `collapisbleContentView`, that was due because some areas of the
view were under one of the corner areas where the floating button could be
and that area was eating the touches