Add ColumnListsView component (#747)
The idea is to render text in columns, the amount of columns can change
depending on the trait collection horizontal size, but this is left to the parent
view to decide the column behavior.

The list is mean to be delivered in the desired order, which in this case is
alphabetical order from top to bottom.

I created a stack view with dynamic text views to implement this as stack view
can solve the layout quite easily for different columns and it's also aware of
the size of the content. Then since the stack view fills the `ColumnListsView`
the latter view is aware of this content, then the height constraint is not needed
in the demo view.