add `KeyValueCollectionView` component
This component is a way to render key/value data as a table with
different columns that shows this value pairs in their cells.

Internally it's a collection view. Then, in order to test iOS 13+
features, like `UICollectionViewDiffableDataSource` and
`UICollectionViewCompositionalLayout` I created proxy objects that will
implemenet the new features and fallback for older versions of iOS, as
a concept mostly to see how we can use today this iOS 13 only features.