Refactor: key value grid view (#765)
* Move array extensions as internal

They will be used in more places than just the ColumnListView

* Rename component to KeyValueGridView

It's not really a collection view anymore

* Refactor KeyValueGridView using Stack Views instead of a UICollectionView

For the kind of data/layout that is gonna be used in this component, a collection
view introduced too much complexity and boiler plate code, specially since the
data is static and won't dynamically change, also the data will be displayed all at
once so there won't be cell reuse. Finally the view needs to know the size of the
data in the appropriate layout, this would happen by calling delegates and
internal mechanisms of UICollectionViews, which was asynchronous and more
complex than needed for this case.

Then I opted to re-write this in stack views instead, as stack views are a way
simpler tool to solve this static grid layout.

The "cell"s themselves are a vertical stack view, which contains 2 labels. The
labels given the content/style "know" their size => then the cell knows it sice
=> then the grid knows it size.

* Rename KeyValyeGridDemoView and update snapshot

* Fix PR comments