Set default value for `UserInterfaceStyleSupport` as `dynamic`
for iOS 13.

Now that we support iOS 13's dark mode, when including the library in a
new project we expect the dynamic colors to be on, which is something
that people can forget.

For older iOS versions, we still use `forceLight` as default value.

This changes the logic for the default value while still allowing
overrides.