Allow setting icon rendering mode for PrimingView
Before, the default was `.alwaysTemplate` for the PrimingView, which is not
what we want all the time (although it's the most common option).

Then I added an optional parameter to the `PrimingViewModel.Row` to allow to
override the rendering mode, but have `.alwaysTemplate` as the default value.

In this way, we keep the existing behavior and API and provide more customization options