Add an optional button for NotificationCenterHeaderView
this button is for mark all as read, and it will only be shown if the `markAllAsReadTitle`
is set, otherwise we don't show it. I did this in this way to bubble up the logic of showing the button or not, since we may not want this button in every section, or may be we do. In any case we don't need to update FinnUI to change this behavior.