Add "more" button to the notification section header
This button will be used to present a bottom sheet for the save search
settings.

This commit includes 2 new requirements for the notification header

- `includeMoreButton` for the `NotificationCenterHeaderViewModel`
- `notificationCenterView(_:segment:didSelectMoreButtonIn)` delegate
method

This means this is a breaking change for FinnUI.

By default the "more" button will be hidden until the view model
indicates it needs it.