Implement saved searches notification pagination
in NotificationCenter

This is a **breaking change for FinnUI** as I introduce a new protocol
requirement for `NotificationCenterViewDataSource`.

```swift
func notificationCenterView(_ view: NotificationCenterView, fetchNextPageFor segment: Int)
```

Which will be called when the user scrolls down the list of
notifications to fetch the next page if necessary.

This only will apply when the saved searches notifications are rendered
chronologically (this logic is handled by the iOS app).

We also keep track of the new fetch request with
`isFetchingNextPageForSavedSearches` to avoid making request when it's
not needed.