Fix FinniversKit.StepSlider for iOS 14
Inspecting the views in reveal it seems that iOS 14 changes the
implementation of the slider, so the internal thumb view was being
rendered at the bottom.

Then when we added the custom subviews `trackView` and
`activeRangeTrackView` they were put on top of the internal thumb view
in the case of iOS 14.